Distinguish business logic from service logic
